model/requirements.txt
numpy
tensorflow
Flask
opencv-python
Pillow